requirements.txt
requests
jinja2
azure-identity
azure-mgmt-apimanagement
python-dotenv
